+++ ios/Classes/SmileIDPlugin.swift
@@ -75,10 +75,6 @@ public class SmileIDPlugin: NSObject, FlutterPlugin, SmileIDApi {
         )
     }
     
-    func setEnvironment(useSandbox: Bool) {
-        SmileID.setEnvironment(useSandbox: useSandbox)
-    }
-    
     func setCallbackUrl(callbackUrl: String) {
         SmileID.setCallbackUrl(url: URL(string: callbackUrl))
     }