+++ lib/smile_id.dart
@@ -26,10 +26,6 @@ class SmileID {
     platformInterface.initialize(config, useSandbox, enableCrashReporting);
   }
 
-  static void setEnvironment({required bool useSandbox}) {
-    platformInterface.setEnvironment(useSandbox);
-  }
-
   static void setCallbackUrl({required Uri callbackUrl}) {
     platformInterface.setCallbackUrl(callbackUrl.toString());
   }