+++ lib/smileid_messages.g.dart
@@ -2268,28 +2268,6 @@ class SmileIDApi {
     }
   }
 
-  Future<void> setEnvironment(bool useSandbox) async {
-    const String __pigeon_channelName = 'dev.flutter.pigeon.smileid.SmileIDApi.setEnvironment';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-      __pigeon_channelName,
-      pigeonChannelCodec,
-      binaryMessenger: __pigeon_binaryMessenger,
-    );
-    final List<Object?>? __pigeon_replyList =
-        await __pigeon_channel.send(<Object?>[useSandbox]) as List<Object?>?;
-    if (__pigeon_replyList == null) {
-      throw _createConnectionError(__pigeon_channelName);
-    } else if (__pigeon_replyList.length > 1) {
-      throw PlatformException(
-        code: __pigeon_replyList[0]! as String,
-        message: __pigeon_replyList[1] as String?,
-        details: __pigeon_replyList[2],
-      );
-    } else {
-      return;
-    }
-  }
-
   Future<void> setCallbackUrl(String callbackUrl) async {
     const String __pigeon_channelName = 'dev.flutter.pigeon.smileid.SmileIDApi.setCallbackUrl';
     final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(